data/Zomato_Data_Insights/requirements.txt
streamlit
pandas
mysql-connector-python
faker
openpyxl
